v0.7 fix: WebP target detection (Windows), MenuLayer::onAchievements invalid syntax (Linux), iOS HttpClient/WebP exclusions + ARC off + remove missing AppIcon
Windows build fix:
- CMakeLists.txt: vcpkg's libwebp port does NOT consistently export a
  'WebP::WebP' imported target across versions - older ports only define
  'WebP::webp' (lowercase) or no target at all, which causes CMake
  configure to fail with 'Target HieuDash links to: WebP::WebP but the
  target was not found.' Be defensive: probe for any of the known target
  names (WebP::WebP, WebP::webp, webp) and fall back to a plain
  find_library lookup if none exist.

Linux build fix:
- Classes/MenuLayer.cpp:100 had 'CCScene* scene = AchievementsLayer::scene
  ? AchievementsLayer::scene() : nullptr;' which is invalid C++ syntax
  (cannot take address of an overloaded member function without '&').
  gcc on Linux correctly rejects this. Replace with a documented no-op
  since AchievementsLayer has no scene() factory method anyway.
- Classes/AppDelegate.cpp: 'return reinterpret_cast<int>(s_pAppDelegate);'
  loses precision on 64-bit platforms (Linux x86_64, iOS arm64, Win64).
  Cast through intptr_t first to silence the error.

iOS build fix:
- proj.ios_mac/project.yml: exclude cocos2dx/platform/CCImageCommonWebp.cpp
  (depends on libwebp.a which was stripped from the GitHub archive).
- proj.ios_mac/project.yml: exclude extensions/network/HttpClient.cpp
  (depends on libcurl.a which was stripped from the GitHub archive;
  Hieu Dash does not call CCHttpClient).
- proj.ios_mac/project.yml: add CLANG_ENABLE_OBJC_ARC: NO (cocos2d-x 2.2.3
  .mm files use manual retain/release).
- proj.ios_mac/project.yml: remove ASSETCATALOG_COMPILER_APPICON_NAME
  since no Asset Catalog is shipped.

Version bump to v0.7:
- .github/workflows/build-release.yml: HIUE_DASH_VERSION='0.7', default
  workflow_dispatch tag='v0.7'.
- proj.android/AndroidManifest.xml: versionCode=7, versionName=0.7.
- proj.android/res/values/strings.xml: app_name='Hieu Dash'.
- proj.ios_mac/ios/Info.plist: CFBundleShortVersionString=0.7.

+++ Classes/AppDelegate.cpp
@@ -8,6 +8,9 @@
 #include "CocosDenshion.h"
 #include "cocos2d.h"
 
+#include <cstdint>
+#include <cstddef>
+
 USING_NS_CC;
 
 // ============================================================================
@@ -142,7 +145,10 @@ void AppDelegate::checkSound() {
 
 int AppDelegate::get() {
     // Returns the singleton's opaque handle (used by some JNI helpers).
-    return reinterpret_cast<int>(s_pAppDelegate);
+    // Cast through intptr_t first to avoid "cast from pointer to smaller
+    // type loses precision" errors on 64-bit platforms (iOS arm64,
+    // Linux x86_64, Windows x64) where sizeof(int) < sizeof(pointer).
+    return static_cast<int>(reinterpret_cast<intptr_t>(s_pAppDelegate));
 }
 
 bool AppDelegate::getIsIOS() {
